[ALPS03641335] DIP: 32-bit standalone drivers
[Detail] 32-bit projects need standalone drivers

[Solution] enabled

MTK-Commit-Id: 38c7f8bc5e00e28fc471def682d7e5b2f50949f1

CR-Id: ALPS03641335
Feature: Driver

+++ drivers/misc/mediatek/cameraisp/mfb/isp_50/camera_mfb.c
@@ -1198,7 +1198,7 @@ static signed int ConfigMFBHW(MFB_Config *pMfbConfig)
 {
 #ifdef MFB_USE_GCE
 		struct cmdqRecStruct *handle;
-		uint64_t engineFlag = (1L << CMDQ_ENG_MFB);
+		uint64_t engineFlag = (uint64_t)(1LL << CMDQ_ENG_MFB);
 #endif
 
 	if (MFB_DBG_DBGLOG == (MFB_DBG_DBGLOG & MFBInfo.DebugMask)) {
